[HZ-1482] Making LikeFunction thread-safe
