+++ hazelcast/src/main/java/com/hazelcast/sql/impl/expression/string/LikeFunction.java
@@ -16,11 +16,13 @@
 
 package com.hazelcast.sql.impl.expression.string;
 
+import com.hazelcast.jet.datamodel.Tuple2;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
 import com.hazelcast.sql.impl.QueryException;
 import com.hazelcast.sql.impl.SqlDataSerializerHook;
+import com.hazelcast.sql.impl.expression.ConcurrentFixedCapacityCache;
 import com.hazelcast.sql.impl.expression.Expression;
 import com.hazelcast.sql.impl.expression.ExpressionEvalContext;
 import com.hazelcast.sql.impl.expression.TriExpression;
@@ -29,6 +31,7 @@ import com.hazelcast.sql.impl.type.QueryDataType;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 import java.io.IOException;
+import java.io.ObjectInputStream;
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -39,8 +42,9 @@ import static com.hazelcast.sql.impl.expression.string.StringFunctionUtils.asVar
  * LIKE string function.
  */
 public class LikeFunction extends TriExpression<Boolean> implements IdentifiedDataSerializable {
+    private static final int PATTERN_CACHE_SIZE = 100;
 
-    private static final long serialVersionUID = 4157617157954663651L;
+    private static final long serialVersionUID = 2L;
 
     /** Single-symbol wildcard in SQL. */
     private static final char ONE_SQL = '_';
@@ -57,10 +61,9 @@ public class LikeFunction extends TriExpression<Boolean> implements IdentifiedDa
     /** Special characters which require escaping in Java. */
     private static final String ESCAPE_CHARACTERS_JAVA = "[]()|^+*?{}$\\.";
 
-    private final Object mux = new Object();
-
     private boolean negated;
-    private transient volatile State state;
+
+    private transient ConcurrentFixedCapacityCache<Tuple2<String, String>, Pattern> patternCache;
 
     public LikeFunction() {
         // No-op.
@@ -70,13 +73,14 @@ public class LikeFunction extends TriExpression<Boolean> implements IdentifiedDa
         super(source, pattern, escape);
 
         this.negated = negated;
+        this.patternCache = new ConcurrentFixedCapacityCache<>(PATTERN_CACHE_SIZE);
     }
 
     public static LikeFunction create(
-        Expression<?> source,
-        Expression<?> pattern,
-        Expression<?> escape,
-        boolean negated
+            Expression<?> source,
+            Expression<?> pattern,
+            Expression<?> escape,
+            boolean negated
     ) {
         return new LikeFunction(source, pattern, escape, negated);
     }
@@ -108,15 +112,7 @@ public class LikeFunction extends TriExpression<Boolean> implements IdentifiedDa
             escape = null;
         }
 
-        if (state == null) {
-            synchronized (mux) {
-                if (state == null) {
-                    state = new State();
-                }
-            }
-        }
-
-        boolean res = state.like(source, pattern, escape);
+        boolean res = like(source, pattern, escape);
 
         if (negated) {
             res = !res;
@@ -152,6 +148,13 @@ public class LikeFunction extends TriExpression<Boolean> implements IdentifiedDa
         super.readData(in);
 
         negated = in.readBoolean();
+        patternCache = new ConcurrentFixedCapacityCache<>(PATTERN_CACHE_SIZE);
+    }
+
+    private void readObject(ObjectInputStream stream) throws ClassNotFoundException, IOException {
+        stream.defaultReadObject();
+        // The transient fields are not initialized during Java deserialization, so we need to do it manually.
+        patternCache = new ConcurrentFixedCapacityCache<>(PATTERN_CACHE_SIZE);
     }
 
     @Override
@@ -178,104 +181,77 @@ public class LikeFunction extends TriExpression<Boolean> implements IdentifiedDa
         return Objects.hash(super.hashCode(), negated);
     }
 
-    /**
-     * Helper class to execute LIKE function. Caches the last observed pattern to avoid constant re-compilation.
-     */
-    public static class State {
-        private final Object mux = new Object();
+    public boolean like(String source, String pattern, String escape) {
+        Pattern javaPattern = convertToJavaPattern(pattern, escape);
 
-        /** Last observed pattern. */
-        private String lastPattern;
+        Matcher matcher = javaPattern.matcher(source);
 
-        /** Last observed escape. */
-        private String lastEscape;
+        return matcher.matches();
+    }
 
-        /** Last Java pattern. */
-        private Pattern lastJavaPattern;
-
-        public boolean like(String source, String pattern, String escape) {
-            Pattern javaPattern = convertToJavaPattern(pattern, escape);
-
-            Matcher matcher = javaPattern.matcher(source);
-
-            return matcher.matches();
-        }
-
-        private Pattern convertToJavaPattern(String pattern, String escape) {
-            synchronized (mux) {
-                if (Objects.equals(pattern, lastPattern) && Objects.equals(escape, lastEscape)) {
-                    return lastJavaPattern;
-                }
-            }
-
+    private Pattern convertToJavaPattern(String pattern, String escape) {
+        Tuple2<String, String> cacheKey = Tuple2.tuple2(pattern, escape);
+        return patternCache.computeIfAbsent(cacheKey, key -> {
             String javaPatternStr = constructJavaPatternString(pattern, escape);
-            Pattern javaPattern = Pattern.compile(javaPatternStr, Pattern.DOTALL);
+            return Pattern.compile(javaPatternStr, Pattern.DOTALL);
+        });
+    }
 
-            synchronized (mux) {
-                lastPattern = pattern;
-                lastEscape = escape;
-                lastJavaPattern = javaPattern;
+    @SuppressWarnings("checkstyle:CyclomaticComplexity")
+    private static String constructJavaPatternString(String pattern, String escape) {
+        // Get the escape character.
+        Character escapeChar;
+
+        if (escape != null) {
+            if (escape.length() != 1) {
+                throw QueryException.error("ESCAPE parameter must be a single character");
             }
 
-            return javaPattern;
+            escapeChar = escape.charAt(0);
+        } else {
+            escapeChar = null;
         }
 
-        @SuppressWarnings("checkstyle:CyclomaticComplexity")
-        private static String constructJavaPatternString(String pattern, String escape) {
-            // Get the escape character.
-            Character escapeChar;
+        // Main logic.
+        StringBuilder javaPattern = new StringBuilder();
 
-            if (escape != null) {
-                if (escape.length() != 1) {
-                    throw QueryException.error("ESCAPE parameter must be a single character");
-                }
+        int i;
 
-                escapeChar = escape.charAt(0);
-            } else {
-                escapeChar = null;
+        for (i = 0; i < pattern.length(); i++) {
+            char patternChar = pattern.charAt(i);
+
+            // Escape special character as needed.
+            if (ESCAPE_CHARACTERS_JAVA.indexOf(patternChar) >= 0) {
+                javaPattern.append('\\');
             }
 
-            // Main logic.
-            StringBuilder javaPattern = new StringBuilder();
-
-            int i;
-
-            for (i = 0; i < pattern.length(); i++) {
-                char patternChar = pattern.charAt(i);
-
-                // Escape special character as needed.
-                if (ESCAPE_CHARACTERS_JAVA.indexOf(patternChar) >= 0) {
-                    javaPattern.append('\\');
+            if (escapeChar != null && patternChar == escapeChar) {
+                if (i == (pattern.length() - 1)) {
+                    throw escapeWildcardsOnly();
                 }
 
-                if (escapeChar != null && patternChar == escapeChar) {
-                    if (i == (pattern.length() - 1)) {
-                        throw escapeWildcardsOnly();
-                    }
+                char nextPatternChar = pattern.charAt(i + 1);
 
-                    char nextPatternChar = pattern.charAt(i + 1);
+                if ((nextPatternChar == ONE_SQL) || (nextPatternChar == MANY_SQL) || (nextPatternChar == escapeChar)) {
+                    javaPattern.append(nextPatternChar);
 
-                    if ((nextPatternChar == ONE_SQL) || (nextPatternChar == MANY_SQL) || (nextPatternChar == escapeChar)) {
-                        javaPattern.append(nextPatternChar);
-
-                        i++;
-                    } else {
-                        throw escapeWildcardsOnly();
-                    }
-                } else if (patternChar == ONE_SQL) {
-                    javaPattern.append(ONE_JAVA);
-                } else if (patternChar == MANY_SQL) {
-                    javaPattern.append(MANY_JAVA);
+                    i++;
                 } else {
-                    javaPattern.append(patternChar);
+                    throw escapeWildcardsOnly();
                 }
+            } else if (patternChar == ONE_SQL) {
+                javaPattern.append(ONE_JAVA);
+            } else if (patternChar == MANY_SQL) {
+                javaPattern.append(MANY_JAVA);
+            } else {
+                javaPattern.append(patternChar);
             }
-
-            return javaPattern.toString();
         }
 
-        private static QueryException escapeWildcardsOnly() {
-            return QueryException.error("Only '_', '%' and the escape character can be escaped");
-        }
+        return javaPattern.toString();
+    }
+
+    private static QueryException escapeWildcardsOnly() {
+        return QueryException.error("Only '_', '%' and the escape character can be escaped");
     }
 }
