+++ hazelcast/src/test/java/com/hazelcast/sql/impl/expression/ConcurrentFixedCapacityCacheTest.java
@@ -1,22 +1,22 @@
 /*
- * Copyright 2021 Hazelcast Inc.
+ * Copyright (c) 2008-2022, Hazelcast, Inc. All Rights Reserved.
  *
- * Licensed under the Hazelcast Community License (the "License");
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://hazelcast.com/hazelcast-community-license
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OF ANY KIND, either express or implied.
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hazelcast.jet.sql.impl.expression.json;
+
+package com.hazelcast.sql.impl.expression;
 
 import com.hazelcast.function.ConsumerEx;
-import com.hazelcast.jet.sql.impl.expression.json.JsonPathUtil.ConcurrentInitialSetCache;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelJVMTest;
 import com.hazelcast.test.annotation.QuickTest;
@@ -35,12 +35,12 @@ import static org.junit.Assert.assertTrue;
 
 @RunWith(HazelcastParallelClassRunner.class)
 @Category({QuickTest.class, ParallelJVMTest.class})
-public class ConcurrentInitialSetCacheTest {
+public class ConcurrentFixedCapacityCacheTest {
     @Test
     public void when_addingElementsToCacheInSingleThread_then_properSizeAndElements() {
         int capacity = 20;
         int elementsToAdd = 100;
-        ConcurrentInitialSetCache<Integer, Integer> cache = new ConcurrentInitialSetCache<>(capacity);
+        ConcurrentFixedCapacityCache<Integer, Integer> cache = new ConcurrentFixedCapacityCache<>(capacity);
         for (int i = 0; i < elementsToAdd; i++) {
             cache.computeIfAbsent(i, Function.identity());
         }
@@ -56,7 +56,7 @@ public class ConcurrentInitialSetCacheTest {
         int capacity = 20;
         int elementsToAdd = 100;
         int threadCount = 10;
-        ConcurrentInitialSetCache<Integer, Integer> cache = new ConcurrentInitialSetCache<>(capacity);
+        ConcurrentFixedCapacityCache<Integer, Integer> cache = new ConcurrentFixedCapacityCache<>(capacity);
         Runnable runnable = () -> {
             for (int i = 0; i < elementsToAdd; i++) {
                 cache.computeIfAbsent(i, Function.identity());
@@ -77,6 +77,6 @@ public class ConcurrentInitialSetCacheTest {
 
     @Test
     public void when_creatingEmptyCache_then_fail() {
-        assertThrows(IllegalArgumentException.class, () -> new ConcurrentInitialSetCache<>(0));
+        assertThrows(IllegalArgumentException.class, () -> new ConcurrentFixedCapacityCache<>(0));
     }
 }
